Major routing refactoring
Links now work only be using PageView events.
These will then set a pseudo url in the browser.
Reloading this pages doesn't work yet (due to probably some wrong usage
of pux as well as missing data loads when loading a subroute directly)

Not sure if this is the best way of doing it yet.
It has the advantage of every transition being clearly marked in the
code as PageView event.
However it isn't really the HTML way of doing it.